+++ packages/fictoan-react/src/components/Row/Row.tsx
@@ -74,9 +74,9 @@ export const Row = React.forwardRef(
         }
 
         if (retainLayoutAlways) {
-            classNames.push(
-                "retain-layout-on-tablet-landscape retain-layout-on-tablet-portrait retain-layout-on-mobile",
-            );
+            classNames.push("retain-layout-on-tablet-landscape");
+            classNames.push("retain-layout-on-tablet-portrait");
+            classNames.push("retain-layout-on-mobile");
         }
 
         if (allowUltraWide) {